add the commits index
